QuickAdminPanel automatic commit

+++ app/Http/Controllers/Api/V1/Admin/InventoryApiController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api\V1\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Http\Controllers\Traits\MediaUploadingTrait;
 use App\Http\Requests\StoreInventoryRequest;
 use App\Http\Requests\UpdateInventoryRequest;
 use App\Http\Resources\Admin\InventoryResource;
@@ -13,16 +14,22 @@ use Symfony\Component\HttpFoundation\Response;
 
 class InventoryApiController extends Controller
 {
+    use MediaUploadingTrait;
+
     public function index()
     {
-        //abort_if(Gate::denies('inventory_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('inventory_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        return new InventoryResource(Inventory::with(['supplier'])->get());
+        return new InventoryResource(Inventory::with(['supplier', 'product'])->get());
     }
 
     public function store(StoreInventoryRequest $request)
     {
         $inventory = Inventory::create($request->all());
+
+        if ($request->input('po_file', false)) {
+            $inventory->addMedia(storage_path('tmp/uploads/' . basename($request->input('po_file'))))->toMediaCollection('po_file');
+        }
 
         return (new InventoryResource($inventory))
             ->response()
@@ -31,14 +38,25 @@ class InventoryApiController extends Controller
 
     public function show(Inventory $inventory)
     {
-        //abort_if(Gate::denies('inventory_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('inventory_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
-        return new InventoryResource($inventory->load(['supplier']));
+        return new InventoryResource($inventory->load(['supplier', 'product']));
     }
 
     public function update(UpdateInventoryRequest $request, Inventory $inventory)
     {
         $inventory->update($request->all());
+
+        if ($request->input('po_file', false)) {
+            if (! $inventory->po_file || $request->input('po_file') !== $inventory->po_file->file_name) {
+                if ($inventory->po_file) {
+                    $inventory->po_file->delete();
+                }
+                $inventory->addMedia(storage_path('tmp/uploads/' . basename($request->input('po_file'))))->toMediaCollection('po_file');
+            }
+        } elseif ($inventory->po_file) {
+            $inventory->po_file->delete();
+        }
 
         return (new InventoryResource($inventory))
             ->response()
@@ -47,7 +65,7 @@ class InventoryApiController extends Controller
 
     public function destroy(Inventory $inventory)
     {
-        //abort_if(Gate::denies('inventory_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(Gate::denies('inventory_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
 
         $inventory->delete();
